Implement serialisation
